implementing and tests for shared-chunks-plugin (#1)
* Initial implementation and tests for shared-chunks-plugin

CommonsChunkPlugin isn't flexible enough to generate the type of shared
chunks that we want. In particular if you pass it a function for minChunks
to scope the included modules to a particular path, it won't include any
dependencies that exist outside of that path. This can be fixed by specifying
a more complicated predicated that includes all the deps, but this is fragile.

This diff implements a plugin called SharedChunksPlugin which handles this
case automatically. It also excludes any modules that have been included in
previous shared chunks and marks those chunks as dependencies of the shared
chunk that's currently being generated.

In a follow-up diff we'll introduce a separate plugin that generates a manifest
file describing all chunk dependencies so that webapp can load only the
chunks necessary for a particular entry point.

Test Plan: npm test

* stop generating empty chunks

* sort parents based on the order they should be loaded

* fix how chunk deps are computed

* fill out README.md