Extract common modules from async chunks as well (#7)
Summary:
Fixes #4 by adding all async chunks to selected chunks.  This
reduces the size of async chunks significantly.  The only
downside is that if the async chunks contain modules that don't
exist in the any of the entry chunks and that would be extracted
based on the settings for the shared chunk, the shared chunk will
be larger.  Those extra modules in the shared chunk will always
be loaded regardless of whether the async chunk is actually loaded.
We will eventually add a setting to create shared chunks containing
modules that fall into situation in a followup diff.

Test Plan: npm test

Reviewers: jacqueline john

Subscribers: csilers jeff mattdunnrankin